moves settlement repo to internal
